+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/read/SupportsReportStatistics.java
@@ -18,7 +18,6 @@
 package org.apache.spark.sql.connector.read;
 
 import org.apache.spark.annotation.Evolving;
-import org.apache.spark.sql.connector.read.stats.Statistics;
 
 /**
  * A mix in interface for {@link Scan}. Data sources can implement this interface to
